fix bad merge from #1211
