+++ plugins/forc-index/src/ops/forc_index_build.rs
@@ -78,14 +78,15 @@ pub fn init(command: BuildCommand) -> anyhow::Result<()> {
 
     let manifest_schema_file = {
         let workspace_root: std::path::PathBuf =
-            crate::ops::utils::cargo_workspace_root_dir(path.as_path()).unwrap();
+            crate::utils::cargo_workspace_root_dir(path.as_path()).unwrap();
         Path::new(&workspace_root).join(manifest.graphql_schema())
     };
 
     // Rebuild the WASM module even if only the schema has changed.
-    crate::ops::utils::ensure_rebuild_if_schema_changed(
+    crate::utils::ensure_rebuild_if_schema_changed(
         root_dir.as_path(),
         Path::new(manifest_schema_file.as_path()),
+        manifest.execution_source(),
     )?;
 
     // Construct our build command
@@ -187,7 +188,7 @@ pub fn init(command: BuildCommand) -> anyhow::Result<()> {
         let profile = if release { "release" } else { "debug" };
 
         let target_dir: std::path::PathBuf =
-            crate::ops::utils::cargo_target_dir(path.as_path()).unwrap();
+            crate::utils::cargo_target_dir(path.as_path()).unwrap();
 
         let abs_artifact_path = target_dir
             .join(defaults::WASM_TARGET)
